testing/mochitest/tests/Harness_sanity/test_sanityParams.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for exposing test suite information</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script class="testbody" type="text/javascript">
ok(SimpleTest.harnessParameters, "Should have parameters.");
</script>
</body>
</html>
